refactor(app): move the text menu under frontend/retro
