+++ crates/app/src/frontend/retro/menu/item.rs
@@ -1,6 +1,6 @@
+use super::{get_menu_toggle, SubMenu, MAX_MENU_ITEM_CHARS};
 use crate::cmd::{AppCmd, BindCmds, BindTarget};
 use crate::config::AppConfig;
-use crate::menu::{get_menu_toggle, SubMenu, MAX_MENU_ITEM_CHARS};
 use crate::roms::RomsState;
 use crate::video::truncate_text;
 use core::auxiliary::joypad::JoypadButton;
